Working on 1.12.2 backend

+++ shared/main/java/net/ccbluex/liquidbounce/api/minecraft/network/play/server/ISPacketEntity.kt
@@ -14,9 +14,9 @@ interface ISPacketEntity : IPacket
 {
 	val rotating: Boolean
 
-	val posX: Byte
-	val posY: Byte
-	val posZ: Byte
+	val posX: Int // FIXME: Check correction code
+	val posY: Int
+	val posZ: Int
 
 	val onGround: Boolean
 	val yaw: Byte
