+++ shared/main/java/net/ccbluex/liquidbounce/api/minecraft/network/play/server/ISPacketSpawnGlobalEntity.kt
@@ -12,7 +12,7 @@ interface ISPacketSpawnGlobalEntity : IPacket
 {
 	val type: Int
 
-	val x: Int
-	val y: Int
-	val z: Int
+	val x: Double
+	val y: Double
+	val z: Double
 }
